undo over engineered routing
