+++ package.json
@@ -43,7 +43,6 @@
     "prettier": "^2.7.1",
     "typescript": "^4.6.4",
     "vite": "^3.0.0",
-    "vite-plugin-top-level-await": "^1.1.1",
     "vite-tsconfig-paths": "^3.5.0"
   }
 }
