+++ src/constants/index.ts
@@ -1,1 +1,2 @@
 export * from './sizes.constants';
+export * from './routes.constants';
