+++ src/constants/routes.constants.ts
@@ -1,0 +1,4 @@
+export const ROUTES = {
+  home: '/',
+  search: '/search',
+};
