+++ src/navigation/routes/index.ts
@@ -1,0 +1,3 @@
+export * from './routes';
+export * from './protected-routes.routes';
+export * from './unprotected-routes.routes';
