+++ src/navigation/routes/protected-routes.routes.ts
@@ -1,0 +1,5 @@
+import { routes, RouteType } from './routes';
+
+export const protectedRoutes: RouteType[] = routes.filter(
+  route => route.isProtected
+);
