+++ src/navigation/routes/unprotected-routes.routes.ts
@@ -1,0 +1,5 @@
+import { routes, RouteType } from './routes';
+
+export const unprotectedRoutes: RouteType[] = routes.filter(
+  route => !route.isProtected
+);
