+++ src/pages/Search/search.page.tsx
@@ -1,4 +1,6 @@
+import { ROUTES } from '@constants';
+
 export const Search: Core.AppPage = () => <h1>search</h1>;
 
-Search.path = '/search';
+Search.path = ROUTES.search;
 Search.isProtected = true;
